arm7: merge SND_lockChannel into SND_exChannel
